config/v2_3_experimental: copy in from config/
Rename package statements at the beginning of files where appropriate,
and also remove parts of translate.go and translate_test.go that don't
convert from 2_2 to 2_3_experimental.

Update import in config/util/parsingErrors.go to point to the new
location.
