+++ config/v2_3_experimental/cloudinit.go
@@ -14,7 +14,7 @@
 
 // These functions are copied from github.com/coreos/coreos-cloudinit/config.
 
-package config
+package v2_3_experimental
 
 import (
 	"bytes"
