+++ config/v2_3_experimental/config.go
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package config
+package v2_3_experimental
 
 import (
 	"bytes"
